+++ weblate/formats/ttkit.py
@@ -1506,12 +1506,13 @@ class JoomlaFormat(PropertiesBaseFormat):
     autoload: tuple[str, ...] = ("*.ini",)
 
 
-class GWTFormat(StringsFormat):
+class GWTFormat(PropertiesBaseFormat):
     # Translators: File format name
     name = gettext_lazy("GWT properties")
     format_id = "gwt"
     loader = ("properties", "gwtfile")
     empty_file_template = "\n"
+    autoload: tuple[str, ...] = ("*.strings",)
     check_flags = ("auto-java-messageformat",)
     language_format = "linux"
     supports_plural: bool = True